--- Iconic/Iconic.csproj
@@ -1,61 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-        <ContentTargetFolders>.</ContentTargetFolders>
-        <PackageId>Koben.Iconic</PackageId>
-        <Title>Iconic</Title>
-        <Description>Highly configurable icon picker for Umbraco. With Iconic you will be able to create a property editor that can use virtually any font package out there and not only that, you will be able to select your icon from more than one package on the same property.		</Description>
-        <Product>Iconic</Product>
-        <PackageTags>umbraco plugin package icon picker editor backoffice</PackageTags>
-        <RootNamespace>Our.Umbraco.Iconic</RootNamespace>
-		<Version>3.1.0</Version>
-		<Authors>Mario Lopez</Authors>
-		<PackageProjectUrl>https://github.com/KOBENDigital/Iconic</PackageProjectUrl>
-		<PackageIcon>images\iconic_logo.png</PackageIcon>
-		<PackageReadmeFile>README.md</PackageReadmeFile>
-		<RepositoryUrl>https://github.com/KOBENDigital/Iconic</RepositoryUrl>
-		<PackageReleaseNotes></PackageReleaseNotes>
-		<AssemblyName>Our.Umbraco.Iconic</AssemblyName>
-		<GeneratePackageOnBuild>False</GeneratePackageOnBuild>
-		<PackageLicenseExpression>MIT</PackageLicenseExpression>		
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Umbraco.Cms.Web.Website" Version="9.0.1" />
-        <PackageReference Include="Umbraco.Cms.Web.BackOffice" Version="9.0.1" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <Content Include="App_Plugins\Iconic\**\*.*">			
-            <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
-            <CopyToPublishDirectory>Always</CopyToPublishDirectory>
-        </Content>
-        <None Include="build\**\*.*">
-            <Pack>True</Pack>
-            <PackagePath>buildTransitive</PackagePath>
-        </None>
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Include="..\README.md">
-        <Pack>True</Pack>
-        <PackagePath></PackagePath>
-      </None>
-    </ItemGroup>
-
-    <ItemGroup>
-      <Content Update="App_Plugins\Iconic\lang\en-US.xml">
-        <SubType>Designer</SubType>
-      </Content>
-      <Content Update="App_Plugins\Iconic\lang\es-ES.xml">
-        <SubType>Designer</SubType>
-      </Content>
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Update="images\iconic_logo.png">
-        <Pack>True</Pack>
-        <PackagePath>\</PackagePath>
-      </None>
-    </ItemGroup>	
-</Project>
