--- Iconic/Models/SelectedIcon.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Iconic.Models
-{
-    public class SelectedIcon
-    {
-        public Guid PackageId { get; set; }
-        public string Icon { get; set; }
-    }
-}
