+++ Koben.Iconic.Core/Models/SelectedIcon.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Koben.Iconic.Core.Models
+{
+    public class SelectedIcon
+    {
+        public Guid PackageId { get; set; }
+        public string Icon { get; set; }
+    }
+}
