--- Tools/publishNuget.ps1
@@ -1,3 +1,0 @@
-﻿param([string]$apiKey, [string]$package)
-
-dotnet nuget push $package -k $apiKey -s https://api.nuget.org/v3/index.json
