--- UmbracoV9/App_Plugins/Iconic/Styles/iconic.scss
@@ -1,38 +1,0 @@
-﻿    .koben-iconic__dialog {
-    select {
-        margin-top: 10px;
-    }
-    .umb-iconpicker-item a {
-        color: black;
-    }
-}
-
-.koben-iconic__container {
-    .umb-panel-header-icon {
-        border: 1px dashed #bbbabf;
-        font-weight: 700;
-        &:hover {
-            color: #2152a3;
-            border-color: #2152a3;
-        }
-        &:focus {
-            outline-color: #6ab4f0;
-        }
-    }
-    &.preview {
-        .umb-panel-header-icon.preview {
-            eview {
-                margin-top: 20px;
-            }
-        }
-    }
-    .umb-iconpicker-item.selected {
-        border: 1px dashed #808080;
-        background-color: lightyellow;
-    }
-    .group-horizontal {
-        display: flex;
-        & * {
-            margin-right: 1em;
-        }
-    }
