--- UmbracoV9/App_Plugins/Iconic/Views/iconic.dialog.html
@@ -1,52 +1,0 @@
-﻿<div class="koben-iconic__dialog" ng-controller="Koben.Iconic.Dialog.Controller">
-
-    <div class="umb-control-group">
-        <div class="form-search">
-            <i class="icon-search"></i>
-            <input type="text"
-                   style="width: 100%"
-                   ng-model="searchTerm"
-                   class="umb-search-field search-query input-block-level"
-                   localize="placeholder"
-                   placeholder="@placeholders_filter"
-                   umb-auto-focus
-                   no-dirty-check>
-        </div>
-    </div>
-
-
-
-    <div class="umb-control-group" ng-if="packages.length > 1">
-        <select class="input-block-level" ng-model="pckgselected"
-                ng-change="loadPackage(pckgselected)"
-                ng-options="p as p.name for p in packages">
-            <option value=""><localize key="iconicDialog_selectOption">Select an option...</localize></option>
-        </select>
-    </div>
-    <div ng-if="packages.length === 1">
-        <h4>{{pckgselected.name}}</h4>
-        <hr />
-    </div>
-
-    <umb-load-indicator ng-if="loading"></umb-load-indicator>
-
-    <div class="umb-control-group" ng-show="!loading && filtered.length > 0">
-        <ul class="umb-iconpicker" ng-show="pckgselected.extractedStyles">
-            <li ng-class="{selected: icon == model.pickerData.icon}" class="umb-iconpicker-item" ng-repeat="icon in filtered = (pckgselected.extractedStyles | filter: searchTerm) track by $index">
-                <button type="button" title="{{icon}}" ng-click="selectIcon(icon)" prevent-default>
-                    <iconic-icon package="pckgselected" icon="icon"></iconic-icon>
-                </button>
-            </li>
-        </ul>
-    </div>
-
-    <umb-empty-state ng-if="pckgselected && filtered.length === 0 && !loading"
-                     position="center">
-        <localize key="iconicDialog_noIconsFound">No icons were found.</localize>
-    </umb-empty-state>
-
-    <umb-empty-state ng-if="!pckgselected && !loading"
-                     position="center">
-        <localize key="iconicDialog_noPackageSelected">No package selected.</localize>
-    </umb-empty-state>
-</div>
