--- UmbracoV9/App_Plugins/Iconic/Views/iconic.html
@@ -1,30 +1,0 @@
-﻿<div class="koben-iconic__container flex" ng-controller="Koben.Iconic.Controller">
-
-    <div class="umb-panel-header-icon" ng-click="overlay.show = true"
-         ng-class="{'-placeholder': model.value===null ||model.value===''}"
-         title="{{model.value.style}}">
-
-        <localize ng-show="!modelIsValid" key="general_add" class="ng-isolate-scope" style="font-size: 1rem;">Add</localize>
-
-        <iconic-icon ng-if="modelIsValid" package="pckg" icon="model.value.icon">
-        </iconic-icon>
-        <umb-load-indicator ng-if="loading"></umb-load-indicator>
-    </div>
-
-    <div>
-        <div ng-if="modelIsValid"><small><em>{{model.value.icon}}</em></small></div>
-
-        <button type="button" ng-if="modelIsValid"
-                class="umb-node-preview__action umb-node-preview__action--red ng-scope"
-                ng-click="removeIcon()"
-                style="margin-left: 0;"
-                prevent-default>
-           <localize key="general_remove">Remove</localize>
-        </button>
-    </div>
-    <umb-overlay ng-if="overlay.show"
-                 model="overlay"
-                 view="overlay.view"
-                 position="right">
-    </umb-overlay>
-</div>
