--- UmbracoV9/App_Plugins/Iconic/Views/iconic.prevalues.html
@@ -1,23 +1,0 @@
-﻿<section class="control" ng-controller="Koben.Iconic.Prevalues.Packages">
-    <div ui-sortable="sortableOptions" ng-model="model.value">
-        <div ng-repeat="item in model.value">
-            <div class="umb-node-preview">
-                <i class="umb-node-preview__icon icon-navigation handle"></i>
-                <div class="umb-node-preview__content">
-                    <div class="umb-node-preview__name">{{item.name}}</div>
-                    <div class="umb-node-preview__description">
-                        <localize key="iconicConfig_icons">Icons</localize>: {{item.extractedStyles.length}}</div>
-                </div>
-                <div class="umb-node-preview__actions">
-                    <a href class="umb-node-preview__action umb-node-preview__action--green" ng-click="editPackage(item)" prevent-default>Edit</a>
-                    <a href class="umb-node-preview__action umb-node-preview__action--red" ng-click="removeItem($index)" prevent-default>
-                        <localize key="iconicConfig_remove">Remove</localize>
-                    </a>
-                </div>
-            </div>
-        </div>
-    </div>
-    <a href class="umb-node-preview-add" ng-click="createNewPackage()" prevent-default>
-        <localize key="general_add">Add</localize>
-    </a>
-</section>
