--- UmbracoV9/App_Plugins/Iconic/package.manifest
@@ -1,39 +1,0 @@
-﻿{
-  "propertyEditors": [
-    {
-      "alias": "koben.iconic",
-      "name": "Iconic",
-      "icon": "icon-binoculars",
-      "group": "Pickers",
-      "editor": {
-        "view": "~/App_Plugins/Iconic/views/iconic.html",
-        "valueType": "JSON"
-      },
-      "prevalues": {
-        "fields": [
-          {
-            "label": "Packages configuration",
-            "description": "Add the font packages you want to use",
-            "key": "packages",
-            "view": "~/App_Plugins/Iconic/views/iconic.prevalues.html"
-          }
-        ]
-      }
-    }
-  ],
-
-  "javascript": [
-    "~/App_Plugins/Iconic/js/build/findPolyfill.min.js",
-    "~/App_Plugins/Iconic/js/src/iconic.models.js",
-    "~/App_Plugins/Iconic/js/src/iconic.controller.js",
-    "~/App_Plugins/Iconic/js/src/iconic.dialog.controller.js",
-    "~/App_Plugins/Iconic/js/src/iconic.prevalues.controller.js",
-    "~/App_Plugins/Iconic/js/src/iconic.directives.js",
-    "~/App_Plugins/Iconic/js/src/iconic.prevalues.editor.controller.js"
-  ],
-  "css": [
-    "~/App_Plugins/Iconic/styles/iconic.min.css"
-  ]
-
-
-}
