+++ UmbracoV9/UmbracoV9.csproj
@@ -9,7 +9,6 @@
 
     <!-- Force windows to use ICU. Otherwise Windows 10 2019H1+ will do it, but older windows 10 and most if not all winodws servers will run NLS -->
     <ItemGroup>
-        <PackageReference Include="Koben.Iconic" Version="3.1.0" />
         <PackageReference Include="Microsoft.ICU.ICU4C.Runtime" Version="68.2.0.6" />
         <RuntimeHostConfigurationOption Include="System.Globalization.AppLocalIcu" Value="68.2" />
     </ItemGroup>
@@ -19,8 +18,6 @@
         <PackageReference Include="Umbraco.Cms.SqlCe" Version="9.0.1" />
         <PackageReference Include="Umbraco.SqlServerCE" Version="4.0.0.1" />
     </ItemGroup>
-
-    <Import Project="..\Iconic\build\Koben.Iconic.targets" />
 
     <ItemGroup>
         <Content Include="App_Plugins/**" CopyToOutputDirectory="Always" />
@@ -37,6 +34,11 @@
       <Folder Include="App_Plugins\" />
     </ItemGroup>
 
+
+    <ItemGroup>
+      <ProjectReference Include="..\Iconic\Koben.Iconic.csproj" />
+    </ItemGroup>
+
     <PropertyGroup>
         <CopyRazorGenerateFilesToPublishDirectory>true</CopyRazorGenerateFilesToPublishDirectory>
     </PropertyGroup>
@@ -46,5 +48,8 @@
         <RazorCompileOnBuild>false</RazorCompileOnBuild>
         <RazorCompileOnPublish>false</RazorCompileOnPublish>
     </PropertyGroup>
+	<Target Name="PreBuild" BeforeTargets="PreBuildEvent">
+		<Exec Command="xcopy $(SolutionDir)Iconic\App_Plugins\Iconic\*.* $(ProjectDir)App_Plugins\Iconic\ /S /E /Y" />
+	</Target>
 
 </Project>
